+++ bookhub-ui/src/app/services/models/page-response-book-response.ts
@@ -1,6 +1,7 @@
 /* tslint:disable */
 /* eslint-disable */
-import { BookResponse } from '../models/book-response';
+import {BookResponse} from './book-response';
+
 export interface PageResponseBookResponse {
   content?: Array<BookResponse>;
   first?: boolean;
