+++ bookhub-ui/src/app/services/models/page-response-borrowed-book-response.ts
@@ -1,6 +1,7 @@
 /* tslint:disable */
 /* eslint-disable */
-import { BorrowedBookResponse } from '../models/borrowed-book-response';
+import {BorrowedBookResponse} from './borrowed-book-response';
+
 export interface PageResponseBorrowedBookResponse {
   content?: Array<BorrowedBookResponse>;
   first?: boolean;
